Add 'games' provider module
Use this to move all files with GAMES category:

grep -lRF 'override val yamlCategory = YamlCategory.GAMES' \
  core/src/main/kotlin/io/github/serpro69/kfaker/provider/ | \
  xargs -I {} mv {} provider/games/src/main/kotlin/io/github/serpro69/kfaker/games/provider/

Then move a couple more, which have their own primary category,
manually.
