+++ provider/games/build.gradle.kts
@@ -1,0 +1,4 @@
+plugins {
+    `faker-lib-conventions`
+    `faker-provider-conventions`
+}
