+++ provider/games/src/main/kotlin/io/github/serpro69/kfaker/games/provider/HeroesOfTheStorm.kt
@@ -1,7 +1,8 @@
-package io.github.serpro69.kfaker.provider
+package io.github.serpro69.kfaker.games.provider
 
 import io.github.serpro69.kfaker.*
 import io.github.serpro69.kfaker.dictionary.*
+import io.github.serpro69.kfaker.provider.YamlFakeDataProvider
 import io.github.serpro69.kfaker.provider.unique.LocalUniqueDataProvider
 import io.github.serpro69.kfaker.provider.unique.UniqueProviderDelegate
